pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "motionxbot"
version = "2.0.0"
description = "Automation-first Discord bot rewritten in Python."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "discord.py==2.7.1",
  "python-dotenv>=1.0.1",
  "aiohttp>=3.10.0",
]

[project.scripts]
motionxbot = "motionxbot.__main__:main"

[tool.setuptools.packages.find]
include = ["motionxbot*"]
